# src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "API Faculdade",
      "description": "Documentação da API Faculdade"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
      "/posts": {
        "get": {
          "summary": "Lista todos os posts",
          "responses": {
            "200": {
              "description": "Lista de posts"
            }
          }
        },
        "post": {
          "summary": "Cria um novo post",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "author": { "type": "string" }
                },
                "required": ["title", "description", "author"]
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Post criado com sucesso"
            }
          }
        }
      },
      "/posts/{id}": {
        "get": {
          "summary": "Busca um post pelo ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Detalhes do post"
            },
            "404": {
              "description": "Post não encontrado"
            }
          }
        },
        "put": {
          "summary": "Atualiza um post pelo ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "author": { "type": "string" }
                },
                "required": ["title", "description", "author"]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Post atualizado com sucesso"
            },
            "404": {
              "description": "Post não encontrado"
            }
          }
        },
        "delete": {
          "summary": "Exclui um post pelo ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Post excluído com sucesso"
            },
            "404": {
              "description": "Post não encontrado"
            }
          }
        }
      }
    }
  }